# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Auto Repost Cleaner API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Account {
  connected: boolean;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  followersCount?: number | null;
  /** @nullable */
  followingCount?: number | null;
  /** @nullable */
  connectedAt?: string | null;
}

export interface AccountConnect {
  username: string;
  sessionToken: string;
}

export interface DailyCount {
  date: string;
  count: number;
}

export interface Stats {
  totalRemoved: number;
  removedToday: number;
  removedThisWeek: number;
  pendingReposts: number;
  /** @nullable */
  lastScanAt?: string | null;
  automationActive: boolean;
  dailyData: DailyCount[];
}

export type RepostStatus = typeof RepostStatus[keyof typeof RepostStatus];


export const RepostStatus = {
  pending: 'pending',
  removed: 'removed',
  failed: 'failed',
} as const;

export interface Repost {
  id: string;
  videoId: string;
  originalAuthor: string;
  /** @nullable */
  description?: string | null;
  repostedAt: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  status: RepostStatus;
}

export interface ScanResult {
  found: number;
  newFound: number;
  scannedAt: string;
}

export interface RemoveResult {
  removed: number;
  failed: number;
}

export interface Automation {
  active: boolean;
  intervalMinutes: number;
  /** @nullable */
  nextRunAt?: string | null;
  /** @nullable */
  startedAt?: string | null;
}

export interface Settings {
  scanIntervalMinutes: number;
  autoRemove: boolean;
  notifications: boolean;
  notifyOnRemoval: boolean;
}

export interface SettingsUpdate {
  scanIntervalMinutes?: number;
  autoRemove?: boolean;
  notifications?: boolean;
  notifyOnRemoval?: boolean;
}

export type LogEntryType = typeof LogEntryType[keyof typeof LogEntryType];


export const LogEntryType = {
  scan: 'scan',
  removal: 'removal',
  automation_start: 'automation_start',
  automation_stop: 'automation_stop',
  account_connect: 'account_connect',
  account_disconnect: 'account_disconnect',
  error: 'error',
} as const;

export interface LogEntry {
  id: string;
  type: LogEntryType;
  message: string;
  /** @nullable */
  detail?: string | null;
  createdAt: string;
}