# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Auto Repost Cleaner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reposts/{id}:
    delete:
      operationId: removeRepost
      tags: [reposts]
      summary: Remove a specific repost by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repost"
...
components:
  schemas:
...
    Repost:
      type: object
      properties:
        id:
          type: string
        videoId:
          type: string
        originalAuthor:
          type: string
        description:
          type: ["string", "null"]
        repostedAt:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, removed, failed]
      required: [id, videoId, originalAuthor, repostedAt, status]
...