# lib/api-zod/src/generated/types/repost.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Auto Repost Cleaner API
 * OpenAPI spec version: 0.1.0
 */
import type { RepostStatus } from './repostStatus';

export interface Repost {
  id: string;
  videoId: string;
  originalAuthor: string;
  /** @nullable */
  description?: string | null;
  repostedAt: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  status: RepostStatus;
}